Fix filtering of outgoing messages
It turns out filtering was completely broken if a bad character was
detected. Instead of filtering out the bad characters, it ended up
filtering the entire message down to "".

Remove \001 from the bad characters list. I have no idea how that got in
there, but \001 is used for CTCP and so obviously is valid.